Add files via upload
This project shows how we can create a tic-tac-toe game using React.js. This project works with this logic: it checks whether there is a winner or not, looks to see if an X or O is suddenly drawn on an occupied field, as this would violate the rules of the game, and writes who won the game, X or O. I should note that the CSS code in this project is not mine, but belongs to artificial intelligence (Gemini).